Prototype for "View Meeting" page (#6)
* Add prototype for view meeting page with voting for time

Created meetings can be viewed in this format.
Includes functionality to vote for a desired meeting time. Note that for the purpose of this prototype, the meeting times are hardcoded. In MVP, meeting times will be fetched from datastore.

* Add hardcoded guest list

* Refactor voteTime to remove repeated code

Loop through all inputs in the vote form instead of repeating the actions for each input separately.`

* Add functionality for user to propose additional times

Added additional input field to allow user to input another time, and add the time to the vote form.
Note that the datetime string is not formatted for the purposes of the prototype.

* Rename script.js to viewMeeting.js

viewMeeting.js contains the JS code for the view-meeting.html page.
Keeping a file name script.js since index.html and home.html depends on it.

* Rename PROPOSED_MEETING_TIMES to 'proposedMeetingTimes'

Variable was not a constant, since it would be incremented.

* Separate element selected by ID into variable

This makes it more clear that the innerText of the element is being updated (increment the number displayed by 1)

* Change <th> to <td>

<th> is reserved for table headers.