validateDuration function and tests
validateDuration() checks that the user inputted hours and minutes for the meeting is valid (meeting is at least 1 minutes long, and no more than 23 hours and 59 minutes (1 day).
Although the input type on the HTML page is a 'range' with restrictions on the min and max values, users can still type in a number that exceeds the max or is less than the min, so values must be checked.
Note that users cannot input non-numeric values, so that edge case is handled.

Also added tests for validateDuration().