Link '+' button to addTimeInput function
addTimeInput will add a datetime-local input field to the page with a unique name.
Currently the uniqueness of the name is maintained by adding a number to the end of a common prefix, and incrementing that integer everytime the user clicks the '+' button.
Note that the <input> and <label>s are enclosed in a div for styling purposes.