Prevent users from proposing meetings times before current date
Set 'min' property of the default meeting time input and all other inputs added by 'addTimeInput' to the ISO date string of current date and time. This will prevent users from entering dates and times that are before current date, and hence cannot schedule meetings to occur in the past.