Verify user inputted a time that is later than current time
Although the datetime-local input 'min' property has been set to the current time, the user can still type in a value OR select a time (hours and minutes) via the calendar input that is in the past, hence this additional check is necessary.
Also added tests for the new rectifyInputtedTime function, and updated tests to addTimeInput() to recognise this addition.