Create initMap() to call createMap() and fetchLocations().
I did this so I could seperate the creation of the map and the adding of the locations to the map. This makes it easier to test as I will only have to test for one thing. Previously, createMap() called fetchLocations(). This meant that to test the creation of the map object I also had to mock the APIs call/fetch involved with the adding of locations to the map.