Create meeting-location-dao to wrap fetch calls to location-data servlets.
Previously, I used a general wrapper class for the fetch functions. I made this change to be more uniform with the other features. This makes the process of the creating the meeting entity uniform as it will use the meeting-locaation-dao and the meeting-time-dao. I chose to make the wrapper functions for cleaner code as this avoids having to call the FetchWrapper constructor before using a method that uses it.