getLoggedInUser() and fetchAndProcess(meetingTimeIds) functions
The fetchAndProcess(meetingTimeIds) will fetch meeting time entities and process them to prepare for dislaying them into a form for users to vote on the meeting times. This function will call getLoggedInUser to get the logged in user and render the form to match the voting status of that user (i.e. disable buttons if they are not allowed to vote, either because they have voted for that meeting time before, or because they have voted MAX_VOTES number of times)