Add 'display' boolean parameter to toggleDeleteButtons() function
Previously, this function will go through all the buttons with the class 'delete-time-button' and change the display style to 'none' if it was 'inline' and vice versa. However this was dangerous, as if the button's display was modified beforehand (e.g. set to 'inline' where it was supposed to be none), this function would have done the opposite of what it was supposed to. Hence include the boolean param to be make it certain whether to show or to hide the buttons.