Manual Time Input Functionality (#23)
* Add content for the manual time proposal page

* Link '+' button to addTimeInput function

addTimeInput will add a datetime-local input field to the page with a unique name.
Currently the uniqueness of the name is maintained by adding a number to the end of a common prefix, and incrementing that integer everytime the user clicks the '+' button.
Note that the <input> and <label>s are enclosed in a div for styling purposes.

* Add corresponding delete buttons for each time input

The delete buttons will delete the particular input field that they have been associated with.
Delete buttons will toggle between showing and hidden - they will only be shown when there are at least 2 input fields, and will be hidden when there is only one input field left on the page.

* Add tests for manualTimeProposal.js functions

Dependency inject document to the functions to make testing DOM manipulations easier.

* rename JS files according to style guide

* Prevent users from proposing meetings times before current date

Set 'min' property of the default meeting time input and all other inputs added by 'addTimeInput' to the ISO date string of current date and time. This will prevent users from entering dates and times that are before current date, and hence cannot schedule meetings to occur in the past.

* Verify user inputted a time that is later than current time

Although the datetime-local input 'min' property has been set to the current time, the user can still type in a value OR select a time (hours and minutes) via the calendar input that is in the past, hence this additional check is necessary.
Also added tests for the new rectifyInputtedTime function, and updated tests to addTimeInput() to recognise this addition.

* Ensure the user inputs unique times

Track the times that the user has already inputted in a Set, and rectify any times that are duplicated.
When the user removes an input, the time that was added to the Set should also be removed.
Updated tests to reflect this.

* Add saveTimes() function that saves times and redirects to next page

Also added an empty HTML page for step 3 of the create meeting process.

* Remove line break before 'else' block

* Update comments to explain the removal of timezone and seconds/milliseconds

* Remove UI checks from manual-time-proposal.js tests

The tests are more focused on functionality rather than UI styles, hence the checks here are not as important.
This will make it easier to modify the UI for stylistic purposes later on.

* Restructure rectifyInputtedTime function

Decouple the UI messages and modification from the logic of verifying whether the time is a future time, for easing testing.

* Add 'display' boolean parameter to toggleDeleteButtons() function

Previously, this function will go through all the buttons with the class 'delete-time-button' and change the display style to 'none' if it was 'inline' and vice versa. However this was dangerous, as if the button's display was modified beforehand (e.g. set to 'inline' where it was supposed to be none), this function would have done the opposite of what it was supposed to. Hence include the boolean param to be make it certain whether to show or to hide the buttons.

* Refactor how enclosingDiv is generated

Create a function that will append an array of children to an enclosing <div> element, rather than appending them manually one by one.