Refactor utility functions outside of manual-time-proposal.js
The inputted date verification methods, and the method to get the datetime string of the current time and date will be needed again to verify that the user inputted period-start and period-end times on the create-meeting-step2-gcal.html pages are in the future. Hence restructure and refactor the functions in preparation for this.