Bug fixes from integration of JS DAOs with MeetingTimeServlet
Use convertMapToJson function to convert HashMaps to JSON
The convertToJson function previously used did not convert more complex objects such as HashMaps to JSON (returned null in those cases). Updated the doPost to return a JSON in the form {meetingTimeId: ID} instead of just the ID string for consistency in return values across the doGet and doPost, and for the function to be able to use convertMapToJson.

Do not store entity Key in separate field 'meetingTimeId'
The Key of the entity is already stored in with the entity, and is accessible via entity.getKey() so do not need to duplicate the data in a separate field. Will also not need to handle the case where multiple entities are retrieved from Datastore with the same key, as automatically generated keys are guanranteed to be unique, according to the docs. (whereas storing the key manually in a separate field will mean that potentially a same value could be stored across two different entities)

Use KeyFactory.keyToString and .stringToKey to convert between Key and String
key.toString() was only to be used for debugging purposes, according to the docs.

Updated tests to reflect these changes