Modify regex to only verify the format of the string
Previously, the regex also did incomplete verification on the validity of the date (e.g. would verify that the digits of the month were 0-3, but this allows '15' to be counted as a valid month). However, this was unnecessary, as trying to construct a Date with the string in the next if block will do the necessary validation to the check the valid date.

Also fixed mispelling in comment.