MeetingTime JavaScript Data Access Object (#25)
* fetchMeetingTime function to fetch MeetingTime entity data from servlet

fetchMeetingTime will fetch and return the data from a MeetingTime Datastore entity, via a Java servlet at the '/meeting-time' endpoint.
Added tests for the function.

* newMeetingTime function and tests

newMeetingTime function will talk to the doPost function in the MeetingTimeServlet class to create and store a new MeetingTime entity in Datastore.

* Encode the value in the query string using encodeURIComponent

The values that are part of the field-value pairs in the query string will be encoded when the query string is sent to the servlet (via the use of URLSearchParams class to construct the query string). Hence the tests need to be updated to reflect this.

* Use fetch().then() structure instead of two 'await's

* Modify regex to only verify the format of the string

Previously, the regex also did incomplete verification on the validity of the date (e.g. would verify that the digits of the month were 0-3, but this allows '15' to be counted as a valid month). However, this was unnecessary, as trying to construct a Date with the string in the next if block will do the necessary validation to the check the valid date.

Also fixed mispelling in comment.