MeetingTime servlet (#30)
* doPost function of MeetingTime servlet

doPost will create a new MeetingTime entity with the relevant details, and store to Datastore.
Added basic tests as well for doPost

Store meetingTimeKey as String rather than Key object for more convenient access in client side.

* MeetingTimeServlet doGet function and tests

Added ServletUtil class to contain utility methods that will be shared by many servlets.

* Change all responses from MeetingTimeServlet to be of JSON type

Originally error responses were sent via response.sendError, which will set the response type to be 'text', which made it diffcult to handle in the front end (would need to account for both JSON and text type responses).

* Add function comments and set content type in sendErrorResponse

* Remove unnecessary line break in MeetingTimeServletTest

* Refactor: move sendErrorResponse() function to ServletUtil class

This function will be useful for all servlets, so moving into this utilities class for it to be callable from other servlets.

* Rename MeetingTimeEntityField.java to MeetingTimeFields

This name was inconsistent with the name of the class inside the file.

* Add helper functions to ServletUtil class and dependencies to pom.xml

* Edit convertMapToJson function to return Json string instead of Json object

* Bug fixes from integration of JS DAOs with MeetingTimeServlet

Use convertMapToJson function to convert HashMaps to JSON
The convertToJson function previously used did not convert more complex objects such as HashMaps to JSON (returned null in those cases). Updated the doPost to return a JSON in the form {meetingTimeId: ID} instead of just the ID string for consistency in return values across the doGet and doPost, and for the function to be able to use convertMapToJson.

Do not store entity Key in separate field 'meetingTimeId'
The Key of the entity is already stored in with the entity, and is accessible via entity.getKey() so do not need to duplicate the data in a separate field. Will also not need to handle the case where multiple entities are retrieved from Datastore with the same key, as automatically generated keys are guanranteed to be unique, according to the docs. (whereas storing the key manually in a separate field will mean that potentially a same value could be stored across two different entities)

Use KeyFactory.keyToString and .stringToKey to convert between Key and String
key.toString() was only to be used for debugging purposes, according to the docs.

Updated tests to reflect these changes

* Fix comment to reflect class function

Co-authored-by: Vivian Ho <vivianho18@outlook.com>