Vote meeting time form display (#46)
* getLoggedInUser() and fetchAndProcess(meetingTimeIds) functions

The fetchAndProcess(meetingTimeIds) will fetch meeting time entities and process them to prepare for dislaying them into a form for users to vote on the meeting times. This function will call getLoggedInUser to get the logged in user and render the form to match the voting status of that user (i.e. disable buttons if they are not allowed to vote, either because they have voted for that meeting time before, or because they have voted MAX_VOTES number of times)

* Implementation for generateVoteTimeForm

This function will be called by fetchAndProcess(meetingTimeIds) to dynamically generate the voting time form given the data passed to it.
Added function voteTime that will update the votecount data in the backend via the VoteMeetingTimeDAO, when the user clicks button to vote for the time.

* Added wrapper function to be called on page load

onload of the page, call displayMeetingTimeForm to first clear the contents of the meeting-times-table div before calling fetchAndProcess(meetingTimeIds) to begin the rendering of the voting form table.

* Handle case where there are no voters yet

time.voters would be null in that case, so there will be no 'includes' property. Without checking for the null case, the program would crash at this line.

* Add 'Vote Time' to innerText of vote buttons

* Update function comments for displayMeetingTimeForm to include param

* Handle error responses from the DAO in voteTime()

use console.error() to log the errors. Not throwing errors (which will stop execution and crash the page, but execution should be able to continue despite the error) and also not alerting the user, since this is a backend error they cannot solve.

* Call displayMeetingTimeForm to re-render the voteform with updated votes

Upon discussion, meetingTimeIds will not be passed to the displayMeetingTimeForm, but can be retrieved by calling a function that @vivianho will write. This will make it possible and easier to re-render the page (no arguments required to displayMeetingTimeForm).

* Refactor error message generation to shared function

* Obtain the meetingTimeIds using MeetingEventDAO

Fetch the meeting event from Datastore via @vivianho's MeetingEventDAO and obtain the meetingTimeIds from the retrieved data.
The meetingEventId is found in the URL of the webpage, in the query string.

* Fix formatting of comments.

* Fix indentation

* Change script tag order to alphabetical

* Remove unnecessary <h2> tag for meeting name

Meeting name will now be displayed in the <h2> with the id "meeting-title"

* Remove duplicate 'dao-utils.js' script tag

* Check lengths of lists match in sortTimeByVotes test + style fixes

There could potentially be the case where both the expected value and actual values are null, and the test would return a false positive. Checking the lengths is additional safeguard against this situation.

* Add comments to explain why sort occurs client-side

* Refactor sorting compare function to use time2.voteCount - time1.voteCount

The return value of the comparison function in the sort of times does not have to be -1, 0, and 1, just needs to be negative, 0 or positive depending on
 whether time2 is smaller, equal or larger than time1 respectively.