AutoProposeTimes class
Class to contain the functions to call the Google Calendar API and to generate the meeting times based on the user data retrieved.
Note that meetingDuration is in milliseconds as dates and times are compared in milliseconds, so the duration in milliseconds will make it more convenient for comparison and date arithmetic later on.