proposeTimes() function with algorithm to auto-generate meeting times
Propose times function uses the following algorithm to analyse a list of TimePeriods representing busy times:
1. Sort the list of TimePeriods in order of earliest to latest start time
2. Scanning over the list of TimePeriod with two  pointers:
    If the end of the current time period (pointed to by left pointer) overlaps with the start of the next time period, move the right pointer one item to the right.
    Otherwise, check the duration of the free time block from the end of the current time period to the start of the next time period.
        If the duration is at least the meeting duration, add to output.
        Move the left pointer to the right pointer (i.e. current is now what the right pointer was pointing to.)

The algorithm time complexity is bounded by the sort of the time periods, O(Nlog(N)), where N is the number of busy TimePeriods in total retrieved from each of the invitees' calendars.

Include the tests for the AutoProposeTime class.