Redirect the create-meeting-step2.html to gcal page accordingly
Add a conditional statement to check whether the user selected 'Google Calendar' or 'Manual', and redirect to the relevant HTML page.
Note that the code assumes that all values of 'method' that is NOT 'gcal' should be redirected to the manual time input page. This is to handle cases where there is already a value stored under the key 'time-find-method' in sessionStorage for some reason that is neither 'gcal' nor 'manual', or if there is NO value retrieved from sessionStorage at all. Time finding using Google Calendar has many more dependencies than manual time input (e.g. API key required, communication with servlets), so manual time is the safest redirect in these rare cases.