generateTimes and rectifyTimeBoundaries functions called onclick
Implementation for the functions to be called onclick of the 'Generate Times' on the create-meeting-step2-gcal.html page.
The generateTimes function will contact the servlet via the GoogleCalendarTimesDAO to fetch the auto-generated meeting times.
Meanwhile, the rectifyTimeBoundaries function will ensure that the user inputted period-start and period-end times are valid. (period-start and period-end indicate the period between which to generate meeting times).

Added tests for the two functions.

NOTE: do not await generateTimes() onclick. 'await' can only be used in async functions, and the onclick function is not async.
However, this should not affect any functionality, since there is no concurrently running dependent scripts at the time that generateTimes() will run, so not 'awaiting' the function will not disrupt any other executions.