Add timezone information to the dates and times used in algorithm
Previously, the default timezone (UTC time) was being used, and there was no way to select a different time zone for the algorithm to operate in.
Added support for this functionality, and also set the default timezone to be AET (Australian Eastern Time) to match that of the place the application is being developed at.
This new functionality will allow also allow client code to have control of what timezone the startTime and endTime are assumed to be in (easier for client code to use this class) and allow it to be sure what timezone the returned data will be in.