Add timezone information to the datetimes used in this servlet
When the periodStart and periodEnd times are passed to AutoProposeTimes.proposeTimes(), they will be assumed to be in the timezone as specified.
This will ensure that client code can be sure of what timezone they algorithm and servlet is working in, and also have control over what timezone it wants the servlet and the AutoProposeTimes class to take the given times to be in.
Also return the time strings in RFC3339 format, for consistency of data transfer from backend to front end.