Add HTML page for Google Calendar time proposal selection  (#53)
* Add radio button for Google Calendar option in create-meeting-step2.html

* HTML page for the Google Calendar auto find times option

If the user selects the Google Calendar option in create-meeting-step2.html, they will be redirected to this page to input the start and end time of the period they would like to generate meeting times for.

* Redirect the create-meeting-step2.html to gcal page accordingly

Add a conditional statement to check whether the user selected 'Google Calendar' or 'Manual', and redirect to the relevant HTML page.
Note that the code assumes that all values of 'method' that is NOT 'gcal' should be redirected to the manual time input page. This is to handle cases where there is already a value stored under the key 'time-find-method' in sessionStorage for some reason that is neither 'gcal' nor 'manual', or if there is NO value retrieved from sessionStorage at all. Time finding using Google Calendar has many more dependencies than manual time input (e.g. API key required, communication with servlets), so manual time is the safest redirect in these rare cases.

* Refactor utility functions outside of manual-time-proposal.js

The inputted date verification methods, and the method to get the datetime string of the current time and date will be needed again to verify that the user inputted period-start and period-end times on the create-meeting-step2-gcal.html pages are in the future. Hence restructure and refactor the functions in preparation for this.

* Refactor verification of future time into separate function

The inputs on the create-meeting-step2-gcal.html page will only need to be verified to be in the future, rather than unique. This refactor is in preparation for that.

* generateTimes and rectifyTimeBoundaries functions called onclick

Implementation for the functions to be called onclick of the 'Generate Times' on the create-meeting-step2-gcal.html page.
The generateTimes function will contact the servlet via the GoogleCalendarTimesDAO to fetch the auto-generated meeting times.
Meanwhile, the rectifyTimeBoundaries function will ensure that the user inputted period-start and period-end times are valid. (period-start and period-end indicate the period between which to generate meeting times).

Added tests for the two functions.

NOTE: do not await generateTimes() onclick. 'await' can only be used in async functions, and the onclick function is not async.
However, this should not affect any functionality, since there is no concurrently running dependent scripts at the time that generateTimes() will run, so not 'awaiting' the function will not disrupt any other executions.

* Function displayAutoGeneratedTimes to display the generated meeting times

displayAutoGeneratedTimes will be called after fetching the automatically generated times from the servlet to display the times (if the server response was successful) or the error message and redirect button (if the server response was no OK).

* Save generated times to sessionStorage and redirect user to next step

- Modify displayAutoGeneratedTimes to return the times displayed to the user, or null if the serverResponse was an error
- Pass the returned times from displayAutoGeneratedTimes to setupNextButton(), which sets the onclick property of the Next button to saveAutoGeneratedTimes(times), and redirectTo(href)
- Implementation and tests for the saveAutoGeneratedTimes and redirectTo functions

* Bug fix: onclick property of redirectButton

The 'click' property was previously being set to string, however, instead the 'onclick' property should be set to a function.

* Update comment to refer to timezone in general

* Simplify if-else blocks that return true/false

* Refactor the display of auto-generated times

Refactor the display of the error message to a separate function for readability, and display the datetime strings in locale string format for a neater UI.

* Rename 'rectifyTimeBoundaries to 'verifyTimeBoundary'

The function returns a boolean value, so 'verifyTimeBoundary' is a more suitable name that reflects what the function does.

* Verify that the error code is outside of range 200-299

HTTP success error codes can fall in the range of 200 - 299, and not necessarily just 200, as was the condition before.

* Rename 'verifyUniqueFutureTime' to 'addUniqueFutureTime'

The function modifies state in additing the time the 'enteredTimes' set, so 'addUniqueFutureTime' is more reflective of what the function's purpose is.